refactor(frontend): extract InfoPanelRenderer from MusicGraph (Stage J)
Pulls the five info-panel render methods and their two HTML helpers out
of the 2562-line MusicGraph god class into a dedicated
`frontend/src/visualization/InfoPanelRenderer.js` module.

Methods moved (and renamed without the `_` prefix, since they are now
the renderer's public API):
- renderSongDetails        (was _renderSongDetails)
- renderCurateRow          (was _renderCurateRow)
- renderCurateDetail       (was _renderCurateDetail)
- renderReleaseBundleDetail (was _renderReleaseBundleDetail)
- renderClaimDetail        (was _renderClaimDetail)
- escapeHtml               (was _escapeHtml)
- detailField              (was _detailField)

The renderer takes its four MusicGraph-side dependencies as constructor
callbacks rather than reaching through `this`:

    new InfoPanelRenderer({
      attachNavLinkListeners,   // shared helper still on MusicGraph
                                // (renderGroupDetails / renderPersonDetails
                                //  also use it)
      navigateToRelease,
      selectCurateOperation,
      voteFromDetail,
    })

Stage H's frontend characterization test (10 snapshots / 15 tests) is
the contract here. Updated it to:
  - point at InfoPanelRenderer.js instead of MusicGraph.js
  - drift-guard against the renamed signatures
  - stub `escapeHtml` / `detailField` directly on the test `this`
    and put the four nav callbacks under `this.callbacks.*`

The snapshot files are untouched — every byte of HTML the renderer
emits is identical to the pre-extraction baseline. Verified:
  - frontend snapshot test: 15 passed / 10 snapshots matched
  - full backend suite:     484 passed / 218 skipped / 30 snapshots
                             — identical to baseline

MusicGraph.js: 2562 → 2211 lines (-351). InfoPanelRenderer.js: 399 lines.

The remaining handoff-listed extractions (FavoritesManager,
OverlayPositioner, GraphDataLoader) are intentionally deferred — they
have no characterization-test coverage, so moving them in this commit
would be flying blind. They can land in follow-up commits once a
matching snapshot net exists, which the original Stage H plan also
called out.
